requirements.txt
streamlit
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
pypdf
chromadb
sentence-transformers
duckduckgo-search
python-dotenv